Clean up deeply nested nodes and syntax usage
Now that we are using class syntax for Components, we should encourage
students to define and call new methods rather than nesting more
JSX code in the main render call.

Students aren't familiar with advanced syntax like the ternary operator
and template strings. This change attempts to make the code more
readable for such a person.